Week1/requirements.txt
# Generated by Auto-Detector
beautifulsoup4==4.14.3
jieba==0.42.1
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.1
pandas==2.3.3
pillow==12.1.0
pytesseract==0.3.13
python-docx==1.2.0
requests==2.32.5
scikit-learn==1.8.0
seaborn==0.13.2
spacy==3.8.11
streamlit==1.53.0
textblob==0.19.0
tqdm==4.67.1
wordcloud==1.9.6
